refactor code to use requirejs so we can grow the app in more than one file
