+++ index.html
@@ -57,6 +57,6 @@
       </div>
     </div>
 
-    <script src="js/index.js"></script>
+    <script data-main="js/index" src="js/lib/require.js"></script>
   </body>
 </html>
