+++ js/routines/women/kloss.js
@@ -1,0 +1,240 @@
+define(function() {
+    var exercises = {
+      "stretch": [
+        {
+          "group-title": "stretch",
+          "src": "images/stretch/karlie-stretch-1.gif",
+          "link": "http://www.vogue.com/7748641/karlie-kloss-workout-stretch-warm-up-modelfit/",
+          "instructions": "Engage stomach & press upward. Hold to 5-10 sec. Repeat until timer runs out.",
+          "reps":"10",
+          "dur": 30000
+        },
+        {
+          "group-title": "stretch",
+          "src": "images/stretch/karlie-stretch-2.gif",
+          "link": "http://www.vogue.com/7748641/karlie-kloss-workout-stretch-warm-up-modelfit/",
+          "instructions": "Cross left leg over right knee and press legs towards you. Hold no more than 5-10 sec. Switch sides & repeat.",
+          "dur": 30000
+        },
+        {
+          "group-title": "stretch",
+          "src": "images/stretch/karlie-stretch-3.gif",
+          "link": "http://www.vogue.com/7748641/karlie-kloss-workout-stretch-warm-up-modelfit/",
+          "instructions": "Sit on your legs. Use fingers to tilt head to left side & press down gently. Switch to right side. Repeat.",
+          "dur": 30000
+        },
+        {
+          "group-title": "stretch",
+          "src": "images/stretch/karlie-stretch-4.gif",
+          "link": "http://www.vogue.com/7748641/karlie-kloss-workout-stretch-warm-up-modelfit/",
+          "instructions": "Use right hand to grab left elbow. Keep left arm bent & shoulder down & away from ear. Switch to other side. Repeat.",
+          "dur": 30000
+        },
+        {
+          "group-title": "stretch",
+          "src": "images/stretch/karlie-stretch-5.gif",
+          "link": "http://www.vogue.com/7748641/karlie-kloss-workout-stretch-warm-up-modelfit/",
+          "instructions": "Stand up, lunge gently towards right knee & hold no more than 5-10 sec before switching sides.",
+          "reps": "5",
+          "dur": 30000
+        }
+      ],
+      "arms": [
+        {
+          "group-title": "arms",
+          "src": "images/arms/karlie-arms-1.gif",
+          "link": "http://www.vogue.com/7747841/karlie-kloss-workout-better-arms-modelfit-justin-gelband/",
+          "instructions": "Slightly bend knees. Hold water bottle or 3 lb weight with elbows angled slightly upward. Rotate torso to left as you lift left leg in to your stomach. Repeat 10-12 times before switching sides.",
+          "dur": 120000
+        },
+        {
+          "group-title": "arms",
+          "src": "images/arms/karlie-arms-2.gif",
+          "link": "http://www.vogue.com/7747841/karlie-kloss-workout-better-arms-modelfit-justin-gelband/",
+          "instructions": "Hold water bottle with arms stretched out in front of body. Twist toward the leg as you lift and bend it. Do 12 reps & switch sides.",
+          "reps": "12",
+          "dur": 120000
+        },
+        {
+          "group-title": "arms",
+          "src": "images/arms/karlie-arms-3.gif",
+          "link": "http://www.vogue.com/7747841/karlie-kloss-workout-better-arms-modelfit-justin-gelband/",
+          "instructions": "Start in seated position: hips back & knees bent. Slowly shift weight from one bent leg to other as you move side to side.",
+          "reps": "12",
+          "dur": 120000
+        },
+        {
+          "group-title": "arms",
+          "src": "images/arms/karlie-arms-4.gif",
+          "link": "http://www.vogue.com/7747841/karlie-kloss-workout-better-arms-modelfit-justin-gelband/",
+          "instructions": "Use 2 water bottles & hold elbows bent with hands close to face. Twist torso to extend arm @ diagonal while doing slow, controlled punch & extending leg behind you & placing heel down at same time. Do at least 10 reps before switching sides.",
+          "reps": "10-20",
+          "dur": 120000
+        }
+      ],
+      "core": [
+        {
+          "group-title": "core",
+          "src": "images/core/karlie-core-1.gif",
+          "link": "http://www.vogue.com/7647857/karlie-kloss-workout-flat-abs-justin-gelband-modelfit/",
+          "instructions": "Keep back straight, extend left leg behind you while picking up water bottle with left arm and hold it forward. Slowly pull leg & arm towards core. Return to extended position and repeat for 5-10 reps. Then switch sides.",
+          "reps": "5-15",
+          "dur": 120000
+        },
+        {
+          "group-title": "core",
+          "src": "images/core/karlie-core-2.gif",
+          "link": "http://www.vogue.com/7647857/karlie-kloss-workout-flat-abs-justin-gelband-modelfit/",
+          "instructions": "Start in plank position. Move knee towards stomach in straight line. Return to original position and repeat 5-10 reps. Then switch sides.",
+          "reps": "5-15",
+          "dur": 120000
+        },
+        {
+          "group-title": "core",
+          "src": "images/core/karlie-core-3.gif",
+          "link": "http://www.vogue.com/7647857/karlie-kloss-workout-flat-abs-justin-gelband-modelfit/",
+          "instructions": "This one is similar to the one before, except you change the angle of your torso as you move, pulling your thigh towards your stomach to create a slight twist.",
+          "reps": "5-15",
+          "dur": 120000
+        },
+        {
+          "group-title": "core",
+          "src": "images/core/karlie-core-4.gif",
+          "link": "http://www.vogue.com/7647857/karlie-kloss-workout-flat-abs-justin-gelband-modelfit/",
+          "instructions": "Start in same plank pose as before. Perform series of side-to-side rotations, pulling gone arm up into tight, bent reverse L-shape position. Place arm down between twists & keep movements small.",
+          "reps": "5-15",
+          "dur": 120000
+        },
+        {
+          "group-title": "core",
+          "src": "images/core/karlie-core-5.gif",
+          "link": "http://www.vogue.com/7647857/karlie-kloss-workout-flat-abs-justin-gelband-modelfit/",
+          "instructions": "Start in plank pose. Raise arm straight out & up to shoulder level, engaging core & keeping torso from going forward. You got this.",
+          "reps": "5-15",
+          "dur": 120000
+        }
+      ],
+      "butt": [
+        {
+          "group-title": "butt",
+          "src": "images/butt/karlie-butt-1.gif",
+          "link":"http://www.vogue.com/7694535/karlie-kloss-butt-workout-modelfit-justin-gelband/",
+          "instructions": "Activate back & stomach & slide left leg straight behind you and lift to butt level. Pulse upward twice slowly. The smaller the motion & the more resistance, the better. Repeat 10-20 times. Switch sides.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {"group-title": "butt",
+          "src": "images/butt/karlie-butt-2.gif",
+          "link":"http://www.vogue.com/7694535/karlie-kloss-butt-workout-modelfit-justin-gelband/",
+          "instructions": "Similar to the previous exercise, working leg bent slightly and raised to just below hip level. Engage stomach & thigh & point your toe. Repeat 10-20 times. Switch sides.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {"group-title": "butt",
+          "src": "images/butt/karlie-butt-3.gif",
+          "link":"http://www.vogue.com/7694535/karlie-kloss-butt-workout-modelfit-justin-gelband/",
+          "instructions": "Start in the same position. This time, flex the working leg. Repeat 10-20 times. Switch sides.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {"group-title": "butt",
+          "src": "images/butt/karlie-butt-4.gif",
+          "link":"http://www.vogue.com/7694535/karlie-kloss-butt-workout-modelfit-justin-gelband/",
+          "instructions": "Grab a towel or paper plate. Place ball of right foot on top of towel/bowl. Lean forward 30 degrees and slide towel behind you along the floor until extended straight back while other leg performs slow lunge. Return to starting position & repeat at least 8-10 times before switching sides.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {"group-title": "butt",
+          "src": "images/butt/karlie-butt-5.gif",
+          "link":"http://www.vogue.com/7694535/karlie-kloss-butt-workout-modelfit-justin-gelband/",
+          "instructions": "Place towel/bowl next to you, slide leg to the right and then back again. Repeat at least 8-10 times before switching sides.",
+          "reps": "10-20",
+          "dur": 120000
+        }
+      ],
+      "legs": [
+        {
+          "group-title": "legs",
+          "src": "images/legs/karlie-legs-1.gif",
+          "link": "http://www.vogue.com/7628941/karlie-kloss-legs-workout-justin-gelband-modelfit/",
+          "instructions": "Start in slightly seated position with both legs bent. Extend right leg out straight & lift it upward while pointing toe. Tap toe back down, bring leg back in & return to starting position. Repeat 10-20 times before switching sides.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {
+          "group-title": "legs",
+          "src": "images/legs/karlie-legs-2.gif",
+          "link": "http://www.vogue.com/7628941/karlie-kloss-legs-workout-justin-gelband-modelfit/",
+          "instructions": "Start in seated position. Make sure standing leg is always bent and extend working leg out straight. Lift and bend working leg towards stomach before tapping it back down to starting position. Repeat 10-20 times before switching sides.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {
+          "group-title": "legs",
+          "src": "images/legs/karlie-legs-3.gif",
+          "link": "http://www.vogue.com/7628941/karlie-kloss-legs-workout-justin-gelband-modelfit/",
+          "instructions": "Start both legs bend and torso leaning forward. Move right leg behind at diagonal. Tap toe then move right leg behind left leg @ diagonal. Back to right side @ diagonal and lift leg before going back to starting position. Repeat at least 5-10 times before switching legs.",
+          "reps": "5-20",
+          "dur": 120000
+        },
+        {
+          "group-title": "legs",
+          "src": "images/legs/karlie-legs-4.gif",
+          "link": "http://www.vogue.com/7628941/karlie-kloss-legs-workout-justin-gelband-modelfit/",
+          "instructions": "Standard lunge. Step back with one leg and extend it. Bend slowly up & down. Repeat at least 8-10 times before switching.",
+          "reps": "10-20",
+          "dur": 120000
+        },
+        {
+          "group-title": "legs",
+          "src": "images/legs/karlie-legs-5.gif",
+          "link": "http://www.vogue.com/7628941/karlie-kloss-legs-workout-justin-gelband-modelfit/",
+          "instructions": "Start with legs shoulder-width apart. Pull body down slow & sit as far back so that knees are in line with your ankles. Then raise up and down from original position. Never go below knee level!",
+          "reps": "10-20",
+          "dur": 120000
+        }
+      ]
+    };
+
+    var playlist = [
+      exercises['stretch'][0],
+      exercises['stretch'][1],
+      exercises['stretch'][2],
+      exercises['stretch'][3],
+      exercises['stretch'][4],
+
+      exercises['arms'][0],
+      exercises['arms'][1],
+      exercises['arms'][2],
+      exercises['arms'][3],
+
+      exercises['core'][0],
+      exercises['core'][1],
+      exercises['core'][2],
+      exercises['core'][3],
+      exercises['core'][4],
+
+      exercises['butt'][0],
+      exercises['butt'][1],
+      exercises['butt'][2],
+      exercises['butt'][3],
+      exercises['butt'][4],
+
+      exercises['legs'][0],
+      exercises['legs'][1],
+      exercises['legs'][2],
+      exercises['legs'][3],
+      exercises['legs'][4],
+
+      exercises['stretch'][0],
+      exercises['stretch'][1],
+      exercises['stretch'][2],
+      exercises['stretch'][3],
+      exercises['stretch'][4]
+    ];
+
+    return {
+        exercises: exercises,
+        playlist: playlist
+    };
+});
